Integrated AABBtrees with Physics engine fully. Lifetime object implementation done. ***How the life time object works: Every lifetime object must be initialized with it's max life and a pointer to the desired function. It is the job of the user to age the lifetime object on whenever they want. Once the limit is reached, the expiry function is called.
For implementing self-destruction, I have introduced a status variable in every object. The expiry function for this object must be the killObject function which makes the status as dead.
Once this is done, it is the job of the manager to update the tree and remove all of it's dead objects.
The manager is desired to be completely independent of the lifetime class and hence, the manager must handle dead objects on its own.

***How manager must be implemented:
The update function in manager should---
1. call the tree.removeDead() method and use the returned vector of id's of dead nodes to call its own removeDead() method.
2. updates pertaining to manager
3. tree.update()